+++ json/genres.json
@@ -1,1 +1,1 @@
-[{"genre_name":"Rap"},{"genre_name":"Hiphop"},{"genre_name":"bb"},{"genre_name":"ff"},{"genre_name":"dd"}]
+[{"genre_name":"Rap"},{"genre_name":"Hiphop"},{"genre_name":"bb"},{"genre_name":"ff"},{"genre_name":"dd"},{"genre_name":"ll"},{"genre_name":"ff"}]
